opportunity_job_api/src/apps/jobs/serializers/job_enrollment_serializer.py
from rest_framework.serializers import ModelSerializer

from apps.jobs.models import JobEnrollment

class JobEnrollmentSerializer(ModelSerializer):
    class Meta:
        models = JobEnrollment
        fields = "__all__"
