opportunity_job_api/requirements/common.txt
Django==4.2.6
asgiref==3.7.2
django-cors-headers==4.3.0
django-extensions==3.2.3
django-modeltranslation==0.18.11
djangorestframework==3.14.0
django-filter==23.3
drf-spectacular==0.26.5
psycopg2==2.9.9
python-decouple==3.8
python-decouple==3.8
sqlparse==0.4.4
